source of com.github.howardqueen:jspring:0.0.1

+++ src/main/java/com/jspring/web/RestPage.java
@@ -6,15 +6,15 @@ public class RestPage {
 	/**
 	 * Current page
 	 */
-	public int page;
+	public Integer page;
 	/**
 	 * Current page size
 	 */
-	public int size;
+	public Integer size;
 	/**
 	 * Total number
 	 */
-	public int total;
+	public Long total;
 	/**
 	 * 
 	 */
